change spin in Ising from float to int

+++ src/model.jl
@@ -2,12 +2,12 @@ abstract Model
 
 type Ising <: Model
     lat :: Lattice
-    spins :: Vector{Float64}
+    spins :: Vector{Int}
 
     function Ising(lat::Lattice)
         model = new()
         model.lat = lat
-        model.spins = rand([1.0,-1.0], numsites(lat))
+        model.spins = rand([1,-1], numsites(lat))
         return model
     end
 end
